bot/requirements.txt
python-telegram-bot==21.4
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.2.1
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
aiohttp==3.9.3
redis==5.0.1